implement compression flexibility (attic init -c / --compression zlib|lzma)
I had to do a bit of bit-fiddling to preserve backwards compatibility.
Previous code used just 1 byte to determine encryption type and compression was hardcoded to be zlib.
It uses type bytes 0x00, 0x01 and 0x02 for that.
The record layout was rather fixed and there was no variable length part to add a compression type byte.
So I split that type byte: the upper 4bits are compression (0 means zlib as before), the lower 4 bits are for encryption.