reimplement new 0x03 format using msgpack
this is much easier to maintain (and change, if needed) than all those hardcoded offsets.

note:

when using packb to store a namedtuple's data, just the tuple's elements get stored, in order (not the names).
thus, the overhead is rather small. but we can just recreate the namedtuple from the tuple returned by unpackb.

namedtuples are very efficient and prettier to deal with than tuples. alternatively, a dictionary could be used,
but packb would create more overhead for it as key names and values would be stored.