FsyncWorker also does flush and fadvise now: call it WritebackThread
My beautiful code didn't include flush to start with, and certainly not
fadvise.  Now it needs a new name and verbose comments.  Such is life.

No semantic change here.  I did move the flush into WritebackThread.
However it will happen in exactly the same sequence (and still on the same
thread as the writes, which sounds good).