create a separate requirements file for travis - without pytest-benchmark
pytest-benchmark (and its dependency "statistics") needs py 3.3+.

we do not want to run benchmarks on travis anyway, but even its presence
makes trouble due to this, so we just remove it from there again.

as --benchmark-skip is also removed from tox configuration, this means benchmarks
are run by default when pytest-benchmark is installed.