Revert "use build_py to fix build on RTD", disables docs build
This reverts commit 86487d192a9a5ab7ff4eedb92d793485b4c30268.

We will instead commit the generated `.rst` usage and API files
directly into git. The setup commands remain to generate them when the
usage or API changes, but as things are the hoops required to generate
those RST files are way too complicated to justify simply build docs.

See #384.

Conflicts:
	setup.py