borg prune --save-space, fixes #239
standard pruning:
Per chunk that is deleted, it writes a DELETE tag into the repo (and remembers the segment for later compaction).
When it does a commit (it does that AFTER pruning all archives!), it calls compact_segments, which writes the remaining chunks into new segments.
At the very very end, it deletes the segments it does not need any more. This uses a lot of additional space to free up space - bad if you are low on space.

space-saving pruning:
delete one archive, then commit (and compact), repeat.
this is slower, but uses less additional disk space.

also: added some notes to compact_segments() for future optimization