Implement exclusions using regular expressions
The existing option to exclude files and directories, “--exclude”, is
implemented using fnmatch[1]. fnmatch matches the slash (“/”) with “*”
and thus makes it impossible to write patterns where a directory with
a given name should be excluded at a given location in the directory
hierarchy, but not at any other place. Consider this structure:

  home/
  home/aaa
  home/aaa/.thumbnails
  home/user
  home/user/img
  home/user/img/.thumbnails

The intention is to exclude “.thumbnails” in all home directories while
retaining directories with the same name in all other locations. With
fnmatch the pattern “home/*/.thumbnails” matches
“home/user/img/.thumbnails” too.

The desired exclusion can be implemented using the “--exclude-regex”
option introduced with this change:

  --exclude-regex 'home/[^/]+/\.thumbnails'

This change has been discussed in issue #43.

[1] https://docs.python.org/3/library/fnmatch.html

Signed-off-by: Michael Hanselmann <public@hansmi.ch>