create subcommand: add --cap-user option
This option is available when running borg as root (Linux only).

It causes borg to switch to the given user, while retaining the
`CAP_DAC_READ_SEARCH` capability, which enables it to read all files.

This is convenient when backup up things like `/etc`, without wanting to run
borg with full root privileges.