Ignore empty index file.
Empty index file is most likely a result from an unclean
shutdown in the middle of write, e.g. on ext4 with delayed
allocation enabled (default).
Ignoring such a file would get it recreated by other parts of code,
where as not ignoring it leads to an exception about
not being able to read enough bytes from the index.

this commit fixes #1195

Signed-off-by: Oleg Drokin <green@linuxhacker.ru>