WIP: robin hood hashing
This changes the collision handling strategy in _hashindex.c to
robin-hood hashing. Incidentally this should be mutually compatible
with the old version.
This hasn't been properly tested yet, except by the unittests, but
I'm sharing this early to get some feedback. Some testing and a
before-and-after performance measurement will follow.