Archiver.do_extract: Fix leak of downloaded chunk contents caused by preloading
Include condition that path is non empty after applying strip_components into
filter passed to iter_items.

All filtering of files to extract must be done in the filter callable used in
archive.iter_items because iter_items will preload all chunks used in items
it returns. If they are not actually extracted the accumulate in the
responsed dict.