faster quickstart
move the note about free space after the step by step example. it is
unlikely that users will hit out of space conditions on their first
run, and at the end of the example, they will see the not anyways.

this is to make the documentation less scary for new users and easier
to use.