WIP: move most of benchmark code in C reduce inconsistencies
Iterating in python over a list of strings (for the keys to lookup)
was a probable culprit for the large variations in benchmark results
(+/- 10% difference between the average of 200 runs).

This joins the keys to a single string and moves the iteration and
call to hashindex_get to C. The single string means we scan trough
contiguous memory, minimizing cache misses, and calling straight from
C avoids the Python overhead. This means we can compare performance
changes purely in the hashmap implementation.