allow excluding parent and including child, fixes #2314
This fixes the problem raised by issue #2314 by requiring that each root subtree
be fully traversed.

The problem occurs when a patterns file excludes a parent directory P later in
the file, and earlier in the file a subdirectory S of P is included.  Because a
tree is processed recursively, P is processed before S is, but if P is
excluded, then S used to not even be considered, because we wouldn't recurse
into P.  With this fix, we now recurse into P nonetheless, but don't add P (as
a directory entry) to the archive.