change --checkpoint-interval default from 600s to 1800s, fixes #2841
a user was (rightfully) complaining that borg does a whole lot of
writing to .cache and was also quite inefficient because that took
rather long and occurred too frequently.

this was caused by big indexes/caches and the relatively high default
checkpoint frequency (every 300s) in 1.0.

we already have increased that value to 1800s in master (which soon will
be 1.1), so this changeset just does the same for 1.0.x.