Detect non-upgraded Attic repositories
When opening a repository, always try to read the magic number of the
latest segment and compare it to the Attic segment magic (unless the
repository is opened for upgrading). If an Attic segment is detected,
raise a dedicated exception, telling the user to upgrade the repository
first.

Fixes #1933.

(cherry picked from commit 0943b322e3ee286e7672859fe995506c589bcb4e)