debug dump-repo-objs should not depend on a manifest obj
thus, we should not use the with_repository decorator to set up
the key as it uses Manifest.load() to set up the key.

if the Manifest is lost, that does not work.

but we can use some arbitrary object to set up the key, so just
do that.

(cherry picked from commit 2e1f2e84d7fcc706f00168a4a49b22604597654c)