Clarify explanation of encryption options
Text has been replaced, added or re-organised to try to improve the explanation of encryption options.

Some changes also corrected what appeared to be errors, e.g. marking none, repokey and keyfile as non-backwards-compatible modes in places.

I think I still do not understand authentication without encryption. What is the purpose of it, and how does it work? I have not added any information regarding this option, but it would probably be worthwhile adding something in a future update.