fix backing up FIFOs. fixes #4394
The O_NONBLOCK caused EAGAIN errors borg did not deal with, so we
better block and wait until read data is available from e.g. a fifo.

Fixed the --read-special test so it creates a FIFO, feeds it from a
thread and backs it up with borg. Then extracts and checks if correct
data has been backed up.

Also: renamed flags_follow to flags_special_follow (its only intended
use is on symlinks to special files with --read-special).