Elaborate on "Calculating size" (#4866)
create: tell that "Calculating size" time and space needs are caused by --progress

It took me several days to figure out that `--progress` was the culprit to Borg being SIGKILL'd by the kernel, because of OOM, and because of I enabled progress option. This elaborates on what's the calculation for.