upgrade FreeBSD VM to 11.4, pyinstaller to v4.2, fixes #5688
FreeBSD 10.x is not supported any more, the pkg sources are not
available any more, so our VM needed an upgrade.

so, we'll build borg 1.1.x binaries on FBSD 11.4 in future.

additionally, pyinstaller had some regression that is fixed by v4.2.
in v4.2-maint branch i added the FBSD bootloaders, which were missing
as usual.